requirements.txt
appdirs~=1.4.4
certifi~=2023.7.22
charset-normalizer~=3.2.0
click~=8.1.6
contourpy~=1.1.0
cycler~=0.11.0
distro~=1.8.0
docker-pycreds~=0.4.0
et-xmlfile~=1.1.0
filelock~=3.12.2
fonttools~=4.41.1
gitdb~=4.0.10
GitPython~=3.1.32
idna~=3.4
Jinja2~=3.1.2
joblib~=1.3.1
kiwisolver~=1.4.4
markdown-it-py~=3.0.0
MarkupSafe~=2.1.3
matplotlib~=3.7.2
mdurl~=0.1.2
monai-weekly~=1.3.dev2330
mpmath~=1.3.0
natsort~=8.4.0
networkx~=3.1
nibabel~=5.1.0
numpy~=1.25.1
openpyxl~=3.1.2
packaging~=23.1
pandas~=2.0.3
pathtools~=0.1.2
Pillow~=10.0.0
protobuf~=4.23.4
psutil~=5.9.5
Pygments~=2.15.1
pyparsing~=3.0.9
python-dateutil~=2.8.2
pytz~=2023.3
PyYAML~=6.0.1
requests~=2.31.0
rich~=13.4.2
scikit-build~=0.17.6
scikit-learn~=1.3.0
scipy~=1.11.1
seaborn~=0.12.2
sentry-sdk~=1.28.1
setproctitle~=1.3.2
six~=1.16.0
smmap~=5.0.0
sympy~=1.12
threadpoolctl~=3.2.0
tomli~=2.0.1
torch~=2.0.1
typing_extensions~=4.7.1
tzdata~=2023.3
urllib3~=2.0.4
wandb~=0.15.7
tqdm~=4.65.0
tensorboard~=2.13.0
schemdraw
plotly
kaleido